Migrated ProfileServiceClient to SignalR
