+++ Vapour.Client.ServiceClients.Interfaces/IProfileServiceClient.cs
@@ -8,10 +8,16 @@ namespace Vapour.Client.ServiceClients;
 public interface IProfileServiceClient
 {
     Task Initialize();
+
     ObservableCollection<IProfile> ProfileList { get; }
+
     Task<IProfile> CreateNewProfile();
+
     Task DeleteProfile(Guid id);
+
     Task<IProfile> SaveProfile(IProfile profile);
+
     Task SetProfile(string controllerKey, Guid profileId);
-    void StartWebSocket(Action<ProfileChangedMessage> profileChangedHandler);
+
+    void StartListening(Action<ProfileChangedMessage> profileChangedHandler, CancellationToken ct = default);
 }
