+++ Vapour.Server.Host/Profile/ProfileMessageForwarder.cs
@@ -1,7 +1,4 @@
-﻿using System.Net.WebSockets;
-using System.Text;
-
-using Newtonsoft.Json;
+﻿using Microsoft.AspNetCore.SignalR;
 
 using Vapour.Server.Controller;
 using Vapour.Server.Profile;
@@ -10,54 +7,24 @@ using Vapour.Shared.Configuration.Profiles.Services;
 
 namespace Vapour.Server.Host.Profile;
 
-[Obsolete]
 public sealed class ProfileMessageForwarder : IProfileMessageForwarder
 {
+    private readonly IHubContext<ProfileMessageHub, IProfileMessageClient> _hubContext;
     private readonly IProfilesService _profilesService;
-    private readonly List<WebSocket> _sockets = new();
 
-
-    public ProfileMessageForwarder(IProfilesService profilesService)
+    public ProfileMessageForwarder(IProfilesService profilesService,
+        IHubContext<ProfileMessageHub, IProfileMessageClient> hubContext)
     {
         _profilesService = profilesService;
+        _hubContext = hubContext;
         _profilesService.OnActiveProfileChanged += SendOnActiveProfileChanged;
-    }
-
-    public async Task StartListening(WebSocket newSocket)
-    {
-        _sockets.Add(newSocket);
-
-        await Task.Run(async () =>
-        {
-            byte[] buffer = new byte[1024 * 4];
-            WebSocketReceiveResult result =
-                await newSocket.ReceiveAsync(new ArraySegment<byte>(buffer), CancellationToken.None);
-
-            while (!result.CloseStatus.HasValue)
-            {
-                result = await newSocket.ReceiveAsync(new ArraySegment<byte>(buffer), CancellationToken.None);
-            }
-
-            _sockets.Remove(newSocket);
-            await newSocket.CloseAsync(result.CloseStatus.Value, result.CloseStatusDescription, CancellationToken.None);
-        });
     }
 
     private async void SendOnActiveProfileChanged(object sender, ProfileChangedEventArgs e)
     {
-        foreach (WebSocket socket in _sockets)
+        await _hubContext.Clients.All.ProfileChanged(new ProfileChangedMessage
         {
-            if (socket is { State: WebSocketState.Open })
-            {
-                ArraySegment<byte> data = new(
-                    Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(new ProfileChangedMessage
-                    {
-                        ControllerKey = e.ControllerKey,
-                        OldProfileId = e.OldProfile.Id,
-                        NewProfileId = e.NewProfile.Id
-                    })));
-                await socket.SendAsync(data, WebSocketMessageType.Text, true, CancellationToken.None);
-            }
-        }
+            ControllerKey = e.ControllerKey, OldProfileId = e.OldProfile.Id, NewProfileId = e.NewProfile.Id
+        });
     }
 }
