+++ Vapour.Server.Interfaces/Profile/IProfileMessageClient.cs
@@ -1,6 +1,8 @@
-﻿namespace Vapour.Server.Profile;
+﻿using Vapour.Server.Controller;
+
+namespace Vapour.Server.Profile;
 
 public interface IProfileMessageClient
 {
-    
+    Task ProfileChanged(ProfileChangedMessage message);
 }
