+++ Vapour.Server.Interfaces/Profile/IProfileMessageForwarder.cs
@@ -1,9 +1,5 @@
-﻿using System.Net.WebSockets;
+﻿namespace Vapour.Server.Profile;
 
-namespace Vapour.Server.Profile;
-
-[Obsolete]
 public interface IProfileMessageForwarder
 {
-    Task StartListening(WebSocket newSocket);
 }
